Heranca pedido_livraria

+++ 0x06/pedido_livraria/Pedido.java
@@ -10,7 +10,7 @@ public class Pedido {
         this.itens = itens;
     }
 
-    public int getPercentualDesconto() {
+    public double getPercentualDesconto() {
         return percentualDesconto;
     }
